Insert common problems in HORTON installation
Essentially a guide to fixing path and linking problems when
installing HORTON. There are three parts: finding and linking
python modules, extensions and libraries. It gives common techniques
in finding certain types of files. It isn't too extensive, but
it should give people a solid way of solving some of the problems
they're having while explaining why (very briefly). Provided
the user installed the dependencies correctly, this guide should
cover most of the problems, I think, they might encounter when
installing HORTON.

I think one more section can be added before this one, like
`Trouble building or installing dependencies`, but this seems way
to general for me to write quickly.

I don't know how to include STDOUT in rst, but it would be helpful
to use that instead of the bash form.