Change all pronoun 'one' to pronoun 'you'
One can be used to refer a gender neutral indefinite being, whereas
you refers to specifically the reader. I think in the scientific
parts, the use of 'one' seems alright (tech_ref_gaussian_basis.rst),
but for the instructional or tutorial-ish parts of the documentation,
one should use 'you' instead, because it helps one connect with the
reader (vs. you should use 'you' instead, because it helps you
connect with the reader).

But maybe switching up the pronouns might add variations to the
writing, making it less boring? But so far, the text seems to flow
well enough..