Fix up user_hamiltonian_model.rst for prerelease
Apart from the typical formatting, grammar, wording fixes, major
change that was needed was the referencing API documentation
instead of copying and pasting parts of it.

Couple comments:
    line 30: numbered list of one
    line 49: one on top of formula looks weird
    line 61: how do you reference cext class?
    line 80: does this really work?
    line 124: aufbau occupation number model.. is this a real word?
    line 132: is hubbard only good for restricted? (not documented)
Main thing is that the code snippets seem to be a cutout of the
example, i.e. cutout each part of the example and use variables
instaed of numbers. I'm not sure if this is a good or bad thing,
especially since the example is given at the end, combining all the
code snippets together into a bigger picture. Or maybe it is just
lazy