Fix test that broke due to fixed project_orbitals_ortho
There is no simple way to tell whether the new energy value is any
better than the old one. It just shows that this type of tests has
limited value as a unit test. We should move such tests outside the unit
tests and do regression testing instead. That's far beyond the scope of
this fix.