Switch from pep8 to pycodestyle
The pep8 project has been renamed to pycodestyle. Newer releases will be
made under the new name only, so we need to use the new name.

The list of pycodestyle error codes is no longer maintained on the
HORTON website because these are constantly changing. A link to the
documentation of pycodestyle is added instead.