Replace pep257 by pydocstyle
pep257 was renamed to pydocstyle. Future releases are made under this
new name.

The documentation is updated and a reference is made to the Pydocstyle
website for the complete list of error messages. (There is no point in
maintaining such a list on the HORTON site because it gets constantly
updated.)